devleop 위한 커밋

+++ webBizMateProject/src/main/java/com/bizmate/project/repository/IssuePostRepository.java
@@ -3,5 +3,5 @@ package com.bizmate.project.repository;
 import com.bizmate.project.domain.IssuePost;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface IssuePostRepository extends JpaRepository<IssuePost,Integer> {
+public interface IssuePostRepository extends JpaRepository<IssuePost,Long> {
 }
