+++ webBizMateProject/src/main/java/com/bizmate/project/repository/ProjectBoardRepository.java
@@ -3,5 +3,5 @@ package com.bizmate.project.repository;
 import com.bizmate.project.domain.ProjectBoard;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface ProjectBoardRepository extends JpaRepository<ProjectBoard,Integer> {
+public interface ProjectBoardRepository extends JpaRepository<ProjectBoard,Long> {
 }
